+++ Vernou.Text.Json.Tests/ObjectJsonConverterTest/DeserializeJsonArray/DeserializeJsonArrayBaseToArray.cs
@@ -1,0 +1,4 @@
+﻿namespace Vernou.Text.Json.Tests.ObjectJsonConverterTest.DeserializeJsonArray;
+
+public class DeserializeJsonArrayBaseToArray : DeserializeJsonArrayBase<object[], object[]>
+{ }
