+++ Vernou.Text.Json.Tests/ObjectJsonConverterTest/DeserializeJsonArray/DeserializeJsonArrayBaseToIEnumerable.cs
@@ -1,0 +1,4 @@
+﻿namespace Vernou.Text.Json.Tests.ObjectJsonConverterTest.DeserializeJsonArray;
+
+public class DeserializeJsonArrayBaseToIEnumerable : DeserializeJsonArrayBase<IEnumerable<object>, List<object>>
+{ }
