+++ Vernou.Text.Json.Tests/ObjectJsonConverterTest/DeserializeJsonArray/DeserializeJsonArrayBaseToIList.cs
@@ -1,0 +1,4 @@
+﻿namespace Vernou.Text.Json.Tests.ObjectJsonConverterTest.DeserializeJsonArray;
+
+public class DeserializeJsonArrayBaseToIList : DeserializeJsonArrayBase<IList<object>, List<object>>
+{ }
