+++ Vernou.Text.Json.Tests/ObjectJsonConverterTest/DeserializeJsonArray/DeserializeJsonArrayBaseToList.cs
@@ -1,0 +1,4 @@
+﻿namespace Vernou.Text.Json.Tests.ObjectJsonConverterTest.DeserializeJsonArray;
+
+public class DeserializeJsonArrayBaseToList : DeserializeJsonArrayBase<List<object>, List<object>>
+{ }
