+++ Vernou.Text.Json.Tests/ObjectJsonConverterTest/DeserializeJsonArray/DeserializeJsonArrayBaseToObject.cs
@@ -1,0 +1,4 @@
+﻿namespace Vernou.Text.Json.Tests.ObjectJsonConverterTest.DeserializeJsonArray;
+
+public class DeserializeJsonArrayBaseToObject : DeserializeJsonArrayBase<object, List<object>>
+{ }
