+++ Vernou.Text.Json.Tests/ObjectJsonConverterTest/DeserializeJsonObject/DeserializeJsonObjectToDictionary.cs
@@ -1,0 +1,4 @@
+﻿namespace Vernou.Text.Json.Tests.ObjectJsonConverterTest.DeserializeJsonObject;
+
+public class DeserializeJsonObjectToDictionary : DeserializeJsonObjectBase<Dictionary<string, object>>
+{ }
