+++ Vernou.Text.Json.Tests/ObjectJsonConverterTest/DeserializeJsonObject/DeserializeJsonObjectToIDictionary.cs
@@ -1,0 +1,4 @@
+﻿namespace Vernou.Text.Json.Tests.ObjectJsonConverterTest.DeserializeJsonObject;
+
+public class DeserializeJsonObjectToIDictionary : DeserializeJsonObjectBase<IDictionary<string, object>>
+{ }
