+++ Vernou.Text.Json.Tests/ObjectJsonConverterTest/DeserializeJsonObject/DeserializeJsonObjectToObject.cs
@@ -1,0 +1,4 @@
+﻿namespace Vernou.Text.Json.Tests.ObjectJsonConverterTest.DeserializeJsonObject;
+
+public class DeserializeJsonObjectToObject : DeserializeJsonObjectBase<object>
+{ }
